requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
livekit==1.0.6
livekit-agents==1.0.19
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.19
livekit-plugins-deepgram==1.0.19
livekit-plugins-elevenlabs==1.0.19
livekit-plugins-noise-cancellation==0.2.3
livekit-plugins-openai==1.0.19
livekit-plugins-silero==1.0.19
livekit-plugins-turn-detector==1.0.19
livekit-protocol==1.0.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
onnxruntime==1.21.1
openai==1.77.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
sounddevice==0.5.1
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
types-protobuf==4.25.0.20240417
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
